requirements.txt
acres==0.2.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
asyncio==3.4.3
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configobj==5.0.9
configparser==7.1.0
cron-descriptor==1.4.5
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.17.0
fitz==0.0.1.dev2
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
jdeskew==0.2.3
Jinja2==3.1.5
jiter==0.8.2
kombu==5.4.2
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.61.0
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.1.0
prompt_toolkit==3.0.50
prov==2.0.1
puremagic==1.28
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
PyJWT==2.10.1
PyMuPDF==1.25.3
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyxnat==1.6.3
rdflib==6.3.2
redis==5.2.1
requests==2.32.3
scipy==1.15.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traits==7.0.2
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
